requirements.txt
accelerate
black
celery
coverage
flake8
Flask-JWT-Extended
flask-redis
Flask-Session
google-auth-oauthlib
hiredis
librosa
mmtafrica @ git+https://github.com/edaiofficial/mmtafrica.git@104699c00f3a85e015ce442dc7790420b8e29ee3
ollama
openai-whisper
pipdeptree
pre-commit
psycopg2-binary
pyAudioAnalysis
pydub
pyparsing
python-dotenv
pyttsx3
replicate
requests-mock
sentence-transformers
SpeechRecognition
tenacity
tensorflow-estimator
tensorflow-hub
textblob